+++ src/pages/info.jsx
@@ -1,0 +1,7 @@
+export default function Info() {
+  return (
+    <>
+      <div className="mx-auto max-w-7xl px-6">the info</div>
+    </>
+  );
+}
