--- src/pages/info.tsx
@@ -1,7 +1,0 @@
-export default function Info() {
-  return (
-    <>
-      <div className="mx-auto max-w-7xl px-6">the info</div>
-    </>
-  );
-}
